fixed svd++ bug

+++ toolkits/collaborative_filtering/io.hpp
@@ -659,7 +659,7 @@ void load_matrix_market_matrix(const std::string & filename, int offset, int D){
     nnz = rows * cols;
   }
 
-  if (D != cols)
+  if (D != (int)cols)
     logstream(LOG_FATAL)<<"Wrong matrix size detected, should be " << D << " instead of : " << cols << std::endl;
 
   for (i=0; i<nnz; i++){
